+++ packages/lumx-vue/src/components/image-block/index.ts
@@ -1,0 +1,1 @@
+export { default as ImageBlock, type ImageBlockProps, ImageBlockCaptionPosition } from './ImageBlock';