+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Create the `UserBlock` component
     -   Create the `Mosaic` component
     -   Create the `Tabs` component family (`Tab`, `TabList`, `TabPanel`, `TabProvider`)
+    -   Create the `ImageBlock` component
 -   `@lumx/react`:
     -   Create the `Combobox` component family